New Graph Example
Graph Example Fragment with BGL data. It is scrollable and scalable with
gesture. To use a gesture click on the graph. Hold control down and play
with the two points that pop up. Make sure you are doing the correct
gesture for scaling and zooming. Sometimes you need to click on a Point
or Line to make it zoom. Default is zoomed out from all the data.