Added support for another Graph Library
This is the graph library recommended by Husam. It is MPAndroidChart on
github. So we have two choices for charts. MPAndroid however supports
Pie Chart and ViewGraph doesn't. I did manage to get both scrollable and
scallable with gestures but still wierd and you have to almost start it
exactly on the line to scale properly.